move group tests into its own file
